Use `u64` for block count. (#12)
* Use `u64` for block count.

* Add missing `Debug` for `RCA` struct.

* Add missing `Debug` for `CardStatus`.

* Fix tests.

* Remove unneeded cast.

+++ tests/tests.rs
@@ -26,7 +26,7 @@ struct CidRes {
 struct CsdRes {
     version: u8,
     transfer_rate: u8,
-    blocks: u32,
+    blocks: u64,
     size_bytes: u64,
     read_current_minimum_vdd: CurrentConsumption,
     write_current_minimum_vdd: CurrentConsumption,
